split open and is_writeable and test_utils
