+++ stor/obs.py
@@ -282,11 +282,6 @@ class OBSPath(Path):
             DNAnexusError: A dxpy client error occured.
             RemoteError: A s3 client error occurred.
         """
-        from stor_dx.dx import DXPath
-        if six.PY3 and encoding and encoding not in ('utf-8', 'utf8') and \
-                isinstance(self, DXPath):  # pragma: no cover
-            raise ValueError('For DNAnexus paths in Python 3, encoding is always assumed to be '
-                             'utf-8. Please switch your encoding or Python version')
         return OBSFile(self, mode=mode, encoding=encoding)
 
     def list(self):
@@ -569,25 +564,6 @@ class OBSFile(object):
                 self.flush()
             self._buffer.close()
         self.closed = True
-        # we want to wait_on_close on DXPath only if no error was thrown while writing the file
-        if not any(sys.exc_info()):  # pragma: no cover
-            self._wait_on_close()
-
-    def _wait_on_close(self):
-        from stor_dx.dx import DXPath
-        if isinstance(self._path, DXPath):
-            wait_on_close = stor.settings.get()['dx']['wait_on_close']
-            if wait_on_close:
-                from stor_dx import dx
-                with dx._wrap_dx_calls():
-                    f = dxpy.DXFile(dxid=self._path.canonical_resource,
-                                    project=self._path.canonical_project)
-                    try:
-                        f.wait_on_close(wait_on_close)
-                    # ignore timeout because we want client code to deal with it
-                    except dxpy.DXError as e:  # pragma: no cover
-                        if 'timeout' not in str(e):
-                            raise
 
     def flush(self):
         """Flushes the write buffer to the OBS path (if it exists)"""
